lesson 8, home1

+++ src/lesson8/home1/Adder.java
@@ -1,8 +1,8 @@
 package lesson8.home1;
 
 public class Adder extends Arithmetic{
-    public int add(int a, int b) {
-        int sum = a + b;
+    public long add(int a, int b) {
+        long sum = a + b;
 
         return sum;
     }
